Update usage of Shescape
Update drupal_extension_tasks.js to also quote the user-controlled input
to child_process:exec on line 138. Previously, this input was only
escaped but not used properly. This is because "regular" escaping by
Shescape is insufficient when inlined directly into a call to the exec
or execSync functions of child_process.

An alternative approach is to use `{ interpolation: true }` when
escaping (line 33) and not quoting (neither on line 139 nor on line 63).
However, this is generally discouraged as quoting is less error prone.